+++ swann-herrera/src/test/java/esgi/al/cc1/domain/DummyEvent.java
@@ -1,0 +1,20 @@
+package esgi.al.cc1.domain;
+
+import java.time.ZonedDateTime;
+import java.util.UUID;
+
+public class DummyEvent implements Event {
+
+  @Override
+  public UUID getId() {
+    // TODO Auto-generated method stub
+    return null;
+  }
+
+  @Override
+  public ZonedDateTime getOccurredDate() {
+    // TODO Auto-generated method stub
+    return null;
+  }
+
+}