+++ swann-herrera/src/test/java/esgi/al/cc1/domain/DummyEventListener.java
@@ -1,0 +1,17 @@
+package esgi.al.cc1.domain;
+
+public class DummyEventListener implements Listener<DummyEvent> {
+
+  @Override
+  public void accept(DummyEvent t) {
+    // TODO Auto-generated method stub
+
+  }
+
+  @Override
+  public void onEvent(DummyEvent event) {
+    // TODO Auto-generated method stub
+
+  }
+
+}