+++ changelog.md
@@ -40,6 +40,13 @@ No TDD yet only structures
 - user Service implementation
 - test
 - fake Implementation of repo in test
+
+## 2021-11-06
+
+### Added
+- add event bus system
+
+
 [Unreleased]: 
 <!--
 TODO